+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { AuthModal } from './components/auth'
 import { Header } from './components/header'
 import { Introduction } from './components/introduction'
-import { SongList } from './components/songs'
+import { SongList } from './components/songs/list'
 
 function App() {
    return (
