--- src/components/songs/index.js
@@ -1,2 +1,0 @@
-export { SongItem } from './SongItem'
-export { SongList } from './SongList'
