+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
 import App from './App'
-import { store } from './stores'
+import { store } from './store'
 import './assets/base.css'
 import './assets/main.css'
