+++ src/store/slices/modal.js
